+++ src/components/Message/Message.js
@@ -114,6 +114,20 @@ class Message extends Component {
     onMentionsClick: PropTypes.func,
     /** @see See [Channel Context](https://getstream.github.io/stream-chat-react/#channelcontext) */
     onMentionsHover: PropTypes.func,
+    /**
+     * The handler for click event on the user that posted the message
+     *
+     * @param event Dom click event which triggered handler.
+     * @param user the User object for the corresponding user.
+     */
+    onUserClick: PropTypes.func,
+    /**
+     * The handler for hover events on the user that posted the message
+     *
+     * @param event Dom hover event which triggered handler.
+     * @param user the User object for the corresponding user.
+     */
+    onUserHover: PropTypes.func,
     /** @see See [Channel Context](https://getstream.github.io/stream-chat-react/#channelcontext) */
     openThread: PropTypes.func,
     /** Handler to clear the edit state of message. It is defined in [MessageList](https://getstream.github.io/stream-chat-react/#messagelist) component */
@@ -123,6 +137,11 @@ class Message extends Component {
      * Available props - https://getstream.github.io/stream-chat-react/#messageinput
      * */
     additionalMessageInputProps: PropTypes.object,
+    /**
+     * The component that will be rendered if the message has been deleted.
+     * All props are passed into this component.
+     */
+    MessageDeleted: PropTypes.elementType,
   };
 
   static defaultProps = {
@@ -425,6 +444,22 @@ class Message extends Component {
     onMentionsHover(e, message.mentioned_users);
   };
 
+  onUserClick = (e) => {
+    if (typeof this.props.onUserClick !== 'function') {
+      return;
+    }
+
+    this.props.onUserClick(e, this.props.message.user);
+  };
+
+  onUserHover = (e) => {
+    if (typeof this.props.onUserHover !== 'function') {
+      return;
+    }
+
+    this.props.onUserHover(e, this.props.message.user);
+  };
+
   getMessageActions = () => {
     const { message, messageActions: messageActionsProps } = this.props;
     const { mutes } = this.props.channel.getConfig();
@@ -501,6 +536,8 @@ class Message extends Component {
         channelConfig={config}
         onMentionsClickMessage={this.onMentionsClick}
         onMentionsHoverMessage={this.onMentionsHover}
+        onUserClick={this.onUserClick}
+        onUserHover={this.onUserHover}
       />
     );
   }