+++ src/components/MessageInput/MessageInput.js
@@ -318,10 +318,7 @@ class MessageInput extends PureComponent {
         // TODO: show error to user that they should wait until image is uploaded
         return;
       }
-      const dupe = attachments.filter(
-        (attach) => upload.asset_url === attach.url,
-      );
-      if (dupe.length >= 1) continue;
+
       attachments.push({
         type: 'file',
         asset_url: upload.url,