+++ src/components/MessageList/MessageList.js
@@ -3,7 +3,9 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import deepequal from 'deep-equal';
 
+import Center from './Center';
 import MessageNotification from './MessageNotification';
+import CustomNotification from './CustomNotification';
 import { MESSAGE_ACTIONS } from '../../utils';
 import { withChannelContext, withTranslationContext } from '../../context';
 import { Attachment } from '../Attachment';
@@ -15,23 +17,6 @@ import { LoadingIndicator } from '../Loading';
 import { EventComponent } from '../EventComponent';
 import { DateSeparator } from '../DateSeparator';
 import { KEY_CODES } from '../AutoCompleteTextarea';
-
-/* eslint sonarjs/no-duplicate-string: 0 */
-
-const Center = ({ children }) => (
-  <div className="str-chat__list__center">{children}</div>
-);
-
-const Notification = ({ children, active, type }) => {
-  if (active) {
-    return (
-      <div className={`str-chat__custom-notification notification-${type}`}>
-        {children}
-      </div>
-    );
-  }
-  return null;
-};
 
 /**
  * MessageList - The message list components renders a list of messages. Its a consumer of [Channel Context](https://getstream.github.io/stream-chat-react/#channel)
@@ -256,8 +241,6 @@ class MessageList extends PureComponent {
 
     if (scrollToBottom) {
       this.scrollToBottom();
-      // Scroll further once attachments are laoded.
-      setTimeout(this.scrollToBottom, 100);
 
       // remove the scroll notification if we already scrolled down...
       this.state.newMessagesNotification &&
@@ -445,6 +428,9 @@ class MessageList extends PureComponent {
     // create object with empty array for each message id
     const readData = {};
     for (const message of messages) {
+      if (!message || !message.id) {
+        continue;
+      }
       readData[message.id] = [];
     }
 
@@ -458,13 +444,14 @@ class MessageList extends PureComponent {
           userLastReadMsgId = msg.id;
         }
       }
-      if (userLastReadMsgId != null) {
+      if (userLastReadMsgId) {
         readData[userLastReadMsgId] = [
           ...readData[userLastReadMsgId],
           readState.user,
         ];
       }
     }
+
     return readData;
   };
 
@@ -645,6 +632,14 @@ class MessageList extends PureComponent {
       ? this.props.loadMore(this.props.messageLimit)
       : this.props.loadMore();
 
+  _onMessageLoadCaptured = () => {
+    // A load event (emitted by e.g. an <img>) was captured on a message.
+    // In some cases, the loaded asset is larger than the placeholder, which means we have to scroll down.
+    if (!this.userScrolledUp()) {
+      this.scrollToBottom();
+    }
+  };
+
   // eslint-disable-next-line
   render() {
     let allMessages = [...this.props.messages];
@@ -666,8 +661,12 @@ class MessageList extends PureComponent {
     // sort by date
     allMessages.sort((a, b) => a.created_at - b.created_at);
 
-    // get the readData
-    const readData = this.getReadStates(allMessages);
+    // get the readData, but only for messages submitted by the user themselves
+    const readData = this.getReadStates(
+      allMessages.filter(
+        ({ user }) => user && user.id === this.props.client.userID,
+      ),
+    );
 
     const lastReceivedId = this.getLastReceived(allMessages);
     const elements = [];
@@ -723,6 +722,7 @@ class MessageList extends PureComponent {
             className={`str-chat__li str-chat__li--${groupStyles}`}
             key={message.id || message.created_at}
             ref={this.messageRefs[message.id]}
+            onLoadCapture={this._onMessageLoadCaptured}
           >
             <Message
               client={this.props.client}
@@ -810,17 +810,17 @@ class MessageList extends PureComponent {
 
         <div className="str-chat__list-notifications">
           {this.state.notifications.map((notification) => (
-            <Notification
+            <CustomNotification
               active={true}
               key={notification.id}
               type={notification.type}
             >
               {notification.text}
-            </Notification>
+            </CustomNotification>
           ))}
-          <Notification active={!this.state.online} type="error">
+          <CustomNotification active={!this.state.online} type="error">
             {t('Connection failure, reconnecting now...')}
-          </Notification>
+          </CustomNotification>
 
           <MessageNotification
             showNotification={this.state.newMessagesNotification}