+++ src/components/Window/Window.js
@@ -7,7 +7,6 @@ import { withChannelContext } from '../../context';
  * Window - A UI component for conditionally displaying thread or channel.
  *
  * @example ../../docs/Window.md
- * @extends PureComponent
  */
 
 const Window = ({ thread, hideOnThread, children }) => {